+++ src/stories/components/UserCard/UserCard.tsx
@@ -9,8 +9,8 @@ import { useAuthContext } from '../../../core/context/AuthContext';
 import { useThemeContext } from '../../../core/context/ThemeContext';
 import { ButtonType } from '../../../core/enums/buttonTypeEnum';
 import { capitalizeWordWithoutConvertLowerCase, getCorrectRoleApi } from '../../../core/utils/string';
+import ControlledSwitches from '../../../views/CUAbout/Button/ControlledSwitches/ControlledSwitches';
 import { ENABLE_DISABLE_USER_REQUEST } from '../../containers/Auth/ToggleAccountsActiveFlag/ToggleAccountsActiveFlagAPI';
-import ControlledSwitches from '../Button/ControlledSwitches/ControlledSwitches';
 import { CustomButton } from '../CustomButton/CustomButton';
 import AvatarPlaceholder from '../svg/avatar-placeholder';
 import type { UserDTO } from '../../../core/models/dto/authDTO';
