+++ src/views/CUAbout/Button/SmallButton/SmallButton.tsx
