+++ src/views/CUAbout/CuAboutView.tsx
@@ -8,7 +8,6 @@ import { getMarkdownInformation } from '@/core/businessLogic/coreUnitAbout';
 import { getFTEsFromCoreUnit } from '@/core/businessLogic/coreUnits';
 import { useThemeContext } from '@/core/context/ThemeContext';
 import { toAbsoluteURL } from '@/core/utils/urls';
-import BigButton from '@/stories/components/Button/BigButton/BigButton';
 import CardInfoMember from '@/stories/components/CardInfoMember/CardInfoMember';
 import { CoreUnitSummary } from '@/stories/components/CoreUnitSummary/CoreUnitSummary';
 import MdViewerContainer from '@/stories/components/Markdown/MdViewerContainer';
@@ -16,6 +15,7 @@ import CardExpenses from '@/stories/components/NavigationCard/CardExpenses';
 import CardSomethingWrong from '@/stories/components/NavigationCard/CardSomethingWrong';
 import { SEOHead } from '@/stories/components/SEOHead/SEOHead';
 import TeamMember from '@/stories/components/TeamMember/TeamMember';
+import BigButton from './Button/BigButton/BigButton';
 import RelateMips from './RelateMips/RelateMips';
 import { useCuAboutView } from './useCuAboutView';
 import type { Theme } from '@mui/material';
